requirements.txt
requests
python-dotenv
pytest
cryptography
